requirements.txt
asgiref==3.2.7
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.16.1
botocore==1.19.1
cffi==1.14.2
channels==2.4.0
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
Django==3.0.5
django-el-pagination==3.3.0
django-ipware==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
django-storages==1.10.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
Pillow==7.1.1
pkg-resources==0.0.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-memcached==1.51
pytz==2019.3
s3transfer==0.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.11
zope.interface==5.1.0
